Pair internal/clientapi shuffle test with a split source file
